# Conflicts:
#	pom.xml
#	src/main/java/com/imcode/imcms/config/ApplicationConfig.java
#	src/main/java/com/imcode/imcms/config/DBConfig.java
#	src/main/java/com/imcode/imcms/config/MainConfig.java
#	src/main/java/com/imcode/imcms/config/WebConfig.java
#	src/main/webapp/WEB-INF/jsp/image_archive/pages/add_image.jsp
#	src/main/webapp/WEB-INF/jsp/image_archive/pages/external_files/external_files.jsp
#	src/main/webapp/WEB-INF/jsp/image_archive/pages/fragments/change_data.jsp
#	src/main/webapp/WEB-INF/jsp/image_archive/pages/fragments/jscalendar.jsp
#	src/main/webapp/WEB-INF/jsp/image_archive/pages/image_card/image_card.jsp
#	src/main/webapp/WEB-INF/jsp/image_archive/pages/search_image.jsp
#	src/main/webapp/WEB-INF/jsp/imcms/document_search_terms.jsp
#	src/main/webapp/WEB-INF/tags/imcms/ui/dialog.tag
#	src/main/webapp/WEB-INF/templates/lang/admin/textdoc/imcms_admin_headtag.html

+++ src/main/java/imcode/server/Config.java
@@ -38,7 +38,7 @@ public class Config {
 
     private String sessionCookieDomain;
     private String fileAdminRootPaths;
-    private float indexingSchedulePeriodInMinutes;
+    private long indexingSchedulePeriodInMinutes;
     private String documentPathPrefix;
     private int documentCacheMaxSize = 100;
     private String keyStorePath;
@@ -196,11 +196,11 @@ public class Config {
         this.fileAdminRootPaths = fileAdminRootPaths;
     }
 
-    public float getIndexingSchedulePeriodInMinutes() {
+    public long getIndexingSchedulePeriodInMinutes() {
         return indexingSchedulePeriodInMinutes;
     }
 
-    public void setIndexingSchedulePeriodInMinutes(float indexingSchedulePeriodInMinutes) {
+    public void setIndexingSchedulePeriodInMinutes(long indexingSchedulePeriodInMinutes) {
         this.indexingSchedulePeriodInMinutes = indexingSchedulePeriodInMinutes;
     }
